Defensively copy navigation properties during graph attach
Issue #7119

The fundamental problem here is that Answer.AuthorId is being used as the FK for two different relationships. The navigation properties for the relationships effectively pointed to different objects during fixup, which caused the FK to change value, resulting in an inconsistent graph. This kind of overlapping FKs will only work if great care is taken with the entities assigned and FK values given, which was not the case in the repro code.

All that being said, the framework should be resilient to modification of navigations while attaching, so defensive copy code has been added to ensure this.